+++ pages/pleistocene/index.tsx
@@ -7,14 +7,20 @@ import SeriesHero from "~/components/heroes/SeriesHero"
 import Layout from "~/components/Layout"
 import Head from "~/components/metadata/Head"
 import SocialNav from "~/components/SocialNav"
+import MIKE_KEESEY from "~/schema/MIKE_KEESEY"
 import PLEISTOCENE from "~/schema/PLEISTOCENE"
 const Page: NextPage = () => {
     return (
         <OrderContainer>
-            <Head favIconType="pleistocene" socialImagePath="/pleistocene" subject={PLEISTOCENE} />
-            <Layout headerPromos={["pleistocene"]} footerPromos={["paleocene", "parry-and-carney"]} theme="bw">
+            <Head favIconType="pleistocene" socialImagePath="/pleistocene" subject={PLEISTOCENE} author={MIKE_KEESEY} />
+            <Layout
+                headerPromos={["pleistocene"]}
+                footerPromos={["paleocene", "the-flesh", "parry-and-carney"]}
+                theme="bw"
+            >
                 <SeriesHero
                     ageRecommendation="17+"
+                    author="Mike Keesey"
                     id="pleistocene"
                     storeLinks={[{ type: "gumroad", url: "//gumroad.com/keesey" }]}
                     title="Pleistocene"