+++ src/schema/GAGE_ULLMAN.ts
@@ -1,0 +1,7 @@
+import type { Person } from "schema-dts"
+const GAGE_ULLMAN: Person = {
+    "@id": "mailto:gage.ullman@gmail.com",
+    "@type": "Person",
+    name: "Gage Ullman",
+}
+export default GAGE_ULLMAN