+++ src/schema/THE_FLESH.ts
@@ -1,0 +1,20 @@
+import type { ComicStory } from "schema-dts"
+const THE_FLESH: ComicStory = {
+    "@id": "https://www.keesey-comics.com/the-flesh",
+    "@type": "ComicStory",
+    abstract:
+        "In a quiet neighborhood shaken by the sudden emergence of a grotesque abomination, its insatiable hunger leaving a trail of carnage and terror, a varied group of survivors grapple with fear, mistrust, and their primal instincts for self-preservation. As the creature's rampage unfolds, the unlikely survivors emerge, each with their own motives and secrets. Amidst the turmoil, the survivors are forced to confront the grim reality of a world spiraling into oblivion as the situation escalates, revealing the true nature of humanity in the face of unspeakable horror.",
+    accessMode: "textOnVisual",
+    creator: "http://tmkeesey.net",
+    contentRating: "17+",
+    copyrightYear: [2021, 2022, 2024],
+    description:
+        "From a sleepy suburb, a grotesque abomination emerges, its insatiable hunger leaving a trail of carnage and terror, as unlikely survivors fight to endure in a world overrun by unspeakable horror.",
+    genre: "horror",
+    image: "https://www.keesey-comics.com/the-flesh/social/1200x1200.png",
+    inLanguage: "en",
+    name: "The Flesh",
+    publisher: "https://www.keesey-comics.com",
+    url: "https://www.keesey-comics.com/the-flesh",
+}
+export default THE_FLESH